# docs/openapi/fortuna.json
{
  "components": {
    "schemas": {
...
      "RecordTransactionCommand": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "double",
            "type": "number"
          },
          "categoryId": {
            "format": "uuid",
            "type": "string"
          },
          "counterparty": {
            "nullable": true,
            "type": "string"
          },
          "creditCardId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/TransactionDirection"
          },
          "financialAccountId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "ownerId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
      "RecordTransactionCommandOutput": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "double",
            "type": "number"
          },
          "appliedRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "categoryId": {
            "format": "uuid",
            "type": "string"
          },
          "categoryName": {
            "nullable": true,
            "type": "string"
          },
          "counterpartyId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "counterpartyName": {
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "creditCardId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/TransactionDirection"
          },
          "financialAccountId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isLateArriving": {
            "type": "boolean"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "originalAmount": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "originalCurrencyCode": {
            "nullable": true,
            "type": "string"
          },
          "rateDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "statementClosingDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "statementDueDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "statementId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "statementPeriodEnd": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "statementPeriodStart": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "statementPurchaseTotal": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "statementStatus": {
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TransactionTagOutput"
            },
            "nullable": true,
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "RecordTransactionCommandOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RecordTransactionCommandOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TransactionDirection": {
        "enum": [
          1,
          2
        ],
        "format": "int32",
        "type": "integer"
      },
...
      "TransactionLifecycleCommandOutput": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "TransactionLifecycleCommandOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionLifecycleCommandOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "TransactionOutput": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "double",
            "type": "number"
          },
          "appliedRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "categoryId": {
            "format": "uuid",
            "type": "string"
          },
          "categoryName": {
            "nullable": true,
            "type": "string"
          },
          "counterpartyId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "counterpartyName": {
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "creditCardId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "creditCardName": {
            "nullable": true,
            "type": "string"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/TransactionDirection"
          },
          "financialAccountId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "financialAccountName": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "installmentNumber": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "installmentPlanId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isLateArriving": {
            "type": "boolean"
          },
          "isManuallyCorrected": {
            "type": "boolean"
          },
          "isReconciled": {
            "type": "boolean"
          },
          "isTransfer": {
            "type": "boolean"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "originalAmount": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "originalCurrencyCode": {
            "nullable": true,
            "type": "string"
          },
          "rateDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "sourceType": {
            "$ref": "#/components/schemas/TransactionSourceType"
          },
          "statementId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TransactionLabelOutput"
            },
            "nullable": true,
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "TransactionOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "TransactionSearchOutput": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionOutput"
            },
            "nullable": true,
            "type": "array"
          },
          "pageNumber": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "totalItems": {
            "format": "int32",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "readOnly": true,
            "type": "integer"
          },
          "totals": {
            "$ref": "#/components/schemas/TransactionTotalsOutput"
          }
        },
        "type": "object"
      },
      "TransactionSearchOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionSearchOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TransactionTagOutput": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateTransactionCommand": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "double",
            "type": "number"
          },
          "categoryId": {
            "format": "uuid",
            "type": "string"
          },
          "counterparty": {
            "nullable": true,
            "type": "string"
          },
          "creditCardId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/TransactionDirection"
          },
          "financialAccountId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "ownerId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "UpdateTransactionCommandOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UpdateTransactionCommandOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Enter the Heimdall JWT bearer token.",
        "scheme": "Bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Fortuna API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/transactions": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "From",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "To",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "FinancialAccountId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "CreditCardId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "CategoryId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "TagId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "CounterpartyId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "Direction",
            "schema": {
              "$ref": "#/components/schemas/TransactionDirection"
            }
          },
          {
            "in": "query",
            "name": "MinimumAmount",
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "MaximumAmount",
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "Text",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "IncludeDeleted",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "DisplayCurrencyCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "FigureDate",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "Descending",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "PageNumber",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "PageSize",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSearchOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSearchOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSearchOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecordTransactionCommand"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordTransactionCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordTransactionCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordTransactionCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordTransactionCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecordTransactionCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeDeleted",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCommand"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTransactionCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTransactionCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTransactionCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/hard": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/restore": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionLifecycleCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "security": [
    {}
  ],
...
}